+++ bettersolve-client/src/components/contests/TaskResultStyles.module.css
@@ -36,5 +36,3 @@
     padding-top: 15px;
     padding-bottom: 15px;
 }
-.SearchInput{
-}
