+++ bettersolve-client/src/components/contests/contestsPage/styles/ContestsName.module.css
