+++ bettersolve-client/src/components/contests/standings/worldResult.module.css
@@ -1,0 +1,5 @@
+
+.worldResTable tr:nth-of-type(2)
+{
+    background-color: rgba(211, 154, 255, 0.507);
+}
