--- bettersolve-client/src/components/contests/worldTaskResult/worldResult.module.css
@@ -1,5 +1,0 @@
-
-.worldResTable tr:nth-of-type(2)
-{
-    background-color: rgba(127, 255, 212, 0.407);
-}
